위젯 모듈 패키지 구조 및 네이밍 정리
- now-playing/feed 컴포저블을 NowPlayingContent.kt / FeedContent.kt 로 분리
- 위젯 정의(EpisodiveWidget·EpisodiveWidgetReceiver·EpisodiveWidgetLayout)를 모듈 루트로 이동
- UI 컴포저블은 component/ 패키지로 분리 (core:designsystem 컨벤션 일치)
- 패키지/클래스명을 Episodive* 하우스 컨벤션에 맞춤 (nowplaying→component, NowPlayingWidget→EpisodiveWidget 등)
