+++ feature/widget/src/main/kotlin/io/jacob/episodive/feature/widget/EpisodiveWidgetLayout.kt
@@ -1,4 +1,4 @@
-package io.jacob.episodive.feature.widget.nowplaying
+package io.jacob.episodive.feature.widget
 
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
@@ -24,7 +24,7 @@ internal enum class FeedMode {
  * - STRIP: 3열 4개 / 4열 5개 (한 행, 제목 없음)
  * - GRID : 3열 2×3=6개 / 4열 2×4=8개 (제목 포함)
  */
-internal data class WidgetLayout(
+internal data class EpisodiveWidgetLayout(
     val feedMode: FeedMode,
     val feedCount: Int,
     /** GRID 일 때 행당 열 수. STRIP 은 한 행에 [feedCount] 개를 균등 배치. */
@@ -42,22 +42,22 @@ internal data class WidgetLayout(
         private val HEIGHT_STRIP_MIN = 175.dp
         private val HEIGHT_GRID_MIN = 300.dp
 
-        fun forSize(size: DpSize): WidgetLayout {
+        fun forSize(size: DpSize): EpisodiveWidgetLayout {
             val wide = size.width >= WIDTH_WIDE_MIN
             return when {
                 size.height >= HEIGHT_GRID_MIN -> {
                     val columns = if (wide) 4 else 3
                     // 3열 6 / 4열 8
-                    WidgetLayout(FeedMode.GRID, feedCount = columns * 2, gridColumns = columns)
+                    EpisodiveWidgetLayout(FeedMode.GRID, feedCount = columns * 2, gridColumns = columns)
                 }
 
                 size.height >= HEIGHT_STRIP_MIN -> {
                     val count = if (wide) 5 else 4
                     // 3열 4 / 4열 5
-                    WidgetLayout(FeedMode.STRIP, feedCount = count, gridColumns = count)
+                    EpisodiveWidgetLayout(FeedMode.STRIP, feedCount = count, gridColumns = count)
                 }
 
-                else -> WidgetLayout(FeedMode.NONE, feedCount = 0, gridColumns = 0)
+                else -> EpisodiveWidgetLayout(FeedMode.NONE, feedCount = 0, gridColumns = 0)
             }
         }
     }
