+++ feature/widget/src/main/kotlin/io/jacob/episodive/feature/widget/EpisodiveWidgetReceiver.kt
@@ -1,0 +1,10 @@
+package io.jacob.episodive.feature.widget
+
+import androidx.glance.appwidget.GlanceAppWidget
+import androidx.glance.appwidget.GlanceAppWidgetReceiver
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class EpisodiveWidgetReceiver : GlanceAppWidgetReceiver() {
+    override val glanceAppWidget: GlanceAppWidget = EpisodiveWidget()
+}
