--- feature/widget/src/main/kotlin/io/jacob/episodive/feature/widget/nowplaying/NowPlayingWidgetReceiver.kt
@@ -1,10 +1,0 @@
-package io.jacob.episodive.feature.widget.nowplaying
-
-import androidx.glance.appwidget.GlanceAppWidget
-import androidx.glance.appwidget.GlanceAppWidgetReceiver
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class NowPlayingWidgetReceiver : GlanceAppWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = NowPlayingWidget()
-}
